refactor: change delete location in project and team
