+++ ui/src/components/Projects/Create/style.module.scss
@@ -113,11 +113,7 @@
 .save-button {
   @extend %save-button;
 
-  grid-column: 1;
-
-  @media(max-width: $sm) {
-    grid-column: 1/3;
-  }
+  grid-column: 1/3;
 }
 
 .saving-in {
@@ -129,6 +125,25 @@
   }
 }
 
+.del-container {
+  display: flex;
+  padding: 15px;
+}
+
+.del-button {
+  background: $color-error-bg;
+  border: 1px solid $color-error-border;
+  border-radius: 4px;
+  color: $color-button-text;
+  cursor: pointer;
+  flex-shrink: 0;
+  font-size: 13px;
+  height: 48px;
+  outline: none;
+  padding: 0 2em;
+  transition: all 0.2s;
+  width: 100%;
+}
 
 .add-team {
   flex-direction: row;
